+++ TrueCraft.Core/Networking/Packets/UpdateSignPacket.cs
@@ -30,7 +30,7 @@ namespace TrueCraft.Core.Networking.Packets
         public void WritePacket(IMinecraftStream stream)
         {
             stream.WriteInt32(X);
-            stream.WriteInt32(Y);
+            stream.WriteInt16(Y);
             stream.WriteInt32(Z);
             stream.WriteString(Text[0]);
             stream.WriteString(Text[1]);
